Shape array names now match shadser attribute names
An unnecessary mapping step is no longer needed to go from the name of a Shape array to the name of a Shader attribute.  This simplifies things.  There was never a very good reason to allow the ability to give distinct names for each of these other than grammar and pluralization (attribute "position" was filled by shape array "positions", etc).  Now, the Shape arrays are all contained inside an "arrays" member and are named the same as the attribute (so, "this.arrays.postiion").