Decoupled camera canvas/mouse registration from display loop.
All Parametric sections now correctly react to the mouse.
But, other demos won't work yet, because they're not always supposed to own their own canvas.
Need to get that info through caller of display() for most situations.  Parametric is an exception; at least it can work due to the decoupling.